Fixing problem with strdup()
For some reason cnf->output = strdup(path)
caused the the contents of 'path' to be
concatenated to the end of itself.  For example
if the contents of 'path' was '/tmp/output.txt'
then after the strdup(), the contents would be
'/tmp/output.txt/tmp/output.txt'.  I assume a
memory problem in the CLI parsing functions.